🥵 wip: new prefs.js
